pyproject.toml
[project]
name = "expenses-app"
version = "0.1.0"
description = "An API designed to track your expenses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "mypy>=1.14.0",
    "python-dotenv>=1.0.1",
    "pydantic[email,mypy]>=2.10.4",
    "asyncpg>=0.30.0",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.1",
    "python-multipart>=0.0.20",
    "types-python-jose>=3.3.4.20240106",
    "types-passlib>=1.7.7.20241221",
]
